# openapi/harness-core-v1.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "auto-agent HTTP API",
    "version": "v1",
    "description": "Versioned SaaS HTTP/SSE contract for the routes registered by pkg/server. Tenant and subject ownership are resolved from the authenticated Principal; cross-tenant session access is intentionally reported as 404. This pre-GA v1 contract accepts additive fields and routes, while semantic or breaking changes require a new reviewed contract revision. Operations marked x-harness-contract-first are approved control-plane contracts whose server registration is verified before they can be considered implemented. Optional adapters return the documented 501 error when they are not configured."
  },
  "servers": [{"url": "/"}],
...
  "paths": {
...
    "/v1/admin/notification-targets": {
      "get": {"operationId": "listNotificationTargets", "tags": ["admin"], "summary": "List non-secret notification targets for one tenant", "description": "Tenant operators may omit tenant_id to use their own tenant. Account administrators must provide tenant_id. Target configuration, endpoints, and credentials are never returned.", "security": [{"BearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/NotificationTargetTenantID"}], "responses": {"200": {"description": "non-secret target metadata", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NotificationTargetListResponse"}}}}, "400": {"$ref": "#/components/responses/Error"}, "401": {"$ref": "#/components/responses/Error"}, "403": {"$ref": "#/components/responses/Error"}, "409": {"$ref": "#/components/responses/Error"}, "501": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}, "500": {"$ref": "#/components/responses/Error"}}},
      "post": {"operationId": "createNotificationTarget", "tags": ["admin"], "summary": "Create one tenant-scoped notification target", "description": "config is opaque provider-owned JSON accepted only for creation. It can contain secrets and is never included in any response.", "security": [{"BearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NotificationTargetCreateRequest"}}}}, "responses": {"201": {"description": "created non-secret target metadata", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NotificationTargetView"}}}}, "400": {"$ref": "#/components/responses/Error"}, "401": {"$ref": "#/components/responses/Error"}, "403": {"$ref": "#/components/responses/Error"}, "409": {"$ref": "#/components/responses/Error"}, "501": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}, "500": {"$ref": "#/components/responses/Error"}}},
      "put": {"operationId": "updateNotificationTarget", "tags": ["admin"], "summary": "Update one tenant-scoped notification target with revision fencing", "description": "config is optional only for metadata-preserving updates; when present it replaces the opaque provider configuration. config is never included in a response.", "security": [{"BearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NotificationTargetUpdateRequest"}}}}, "responses": {"200": {"description": "updated non-secret target metadata", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NotificationTargetView"}}}}, "400": {"$ref": "#/components/responses/Error"}, "401": {"$ref": "#/components/responses/Error"}, "403": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "409": {"$ref": "#/components/responses/Error"}, "501": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}, "500": {"$ref": "#/components/responses/Error"}}},
      "delete": {"operationId": "deleteNotificationTarget", "tags": ["admin"], "summary": "Delete one tenant-scoped notification target with revision fencing", "security": [{"BearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NotificationTargetDeleteRequest"}}}}, "responses": {"200": {"description": "target deleted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Status"}}}}, "400": {"$ref": "#/components/responses/Error"}, "401": {"$ref": "#/components/responses/Error"}, "403": {"$ref": "#/components/responses/Error"}, "404": {"$ref": "#/components/responses/Error"}, "409": {"$ref": "#/components/responses/Error"}, "501": {"$ref": "#/components/responses/Error"}, "503": {"$ref": "#/components/responses/Error"}, "500": {"$ref": "#/components/responses/Error"}}}
    },
...
    "/v1/admin/accounts": {"get": {"operationId": "listAccounts", "tags": ["admin"], "security": [{"BearerAuth": []}], "responses": {"200": {"description": "tenant-visible accounts for an account or tenant administrator", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountList"}}}}, "401": {"$ref": "#/components/responses/Error"}, "403": {"$ref": "#/components/responses/Error"}}}, "post": {"operationId": "createAccount", "tags": ["admin"], "security": [{"BearerAuth": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountCreateRequest"}}}}, "responses": {"201": {"description": "account created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountCreateResponse"}}}}, "400": {"$ref": "#/components/responses/Error"}, "401": {"$ref": "#/components/responses/Error"}, "403": {"$ref": "#/components/responses/Error"}}}},
    "/v1/admin/accounts/{account}/password": {"post": {"operationId": "setAccountPassword", "tags": ["admin"], "security": [{"BearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/Account"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PasswordSetRequest"}}}}, "responses": {"200": {"$ref": "#/components/responses/Success"}, "400": {"$ref": "#/components/responses/Error"}, "401": {"$ref": "#/components/responses/Error"}, "403": {"$ref": "#/components/responses/Error"}}}},
    "/v1/admin/accounts/{account}/status": {"post": {"operationId": "setAccountStatus", "tags": ["admin"], "security": [{"BearerAuth": []}], "parameters": [{"$ref": "#/components/parameters/Account"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccountStatusRequest"}}}}, "responses": {"200": {"$ref": "#/components/responses/Success"}, "400": {"$ref": "#/components/responses/Error"}, "401": {"$ref": "#/components/responses/Error"}, "403": {"$ref": "#/components/responses/Error"}}}},
    "/v1/admin/tenants": {"get": {"operationId": "listTenants", "tags": ["admin"], "security": [{"BearerAuth": []}], "responses": {"2XX": {"$ref": "#/components/responses/Success"}, "default": {"$ref": "#/components/responses/Error"}}}, "post": {"operationId": "createTenant", "tags": ["admin"], "security": [{"BearerAuth": []}], "responses": {"2XX": {"$ref": "#/components/responses/Success"}, "default": {"$ref": "#/components/responses/Error"}}}},
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {"type": "http", "scheme": "bearer", "description": "Built-in account token, or a deployment-injected Authenticator. The resolved Principal supplies tenant, subject, scope, grants, and role; handlers enforce ownership and administrator boundaries. Development header fallback is disabled in production."},
      "RunnerBearer": {"type": "http", "scheme": "bearer", "description": "Private runner token; it cannot access administrative runner task routes."}
    },
    "parameters": {
...
      "Account": {"name": "account", "in": "path", "required": true, "schema": {"type": "string"}},
...
      "NotificationTargetTenantID": {"name": "tenant_id", "in": "query", "required": false, "schema": {"type": "string"}, "description": "Optional for a tenant operator, which then uses its own tenant. Required for an account administrator; a tenant operator cannot select another tenant."},
...
    },
    "responses": {
      "Success": {"description": "Implemented successful response; exact payload varies by operation."},
      "Error": {"description": "Uniform JSON error envelope", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
    },
    "schemas": {
      "Error": {"type": "object", "required": ["error"], "properties": {"error": {"type": "string"}}, "additionalProperties": false},
...
      "Status": {"type": "object", "required": ["status"], "properties": {"status": {"type": "string"}}},
      "NotificationTargetView": {"type": "object", "description": "Non-secret notification target metadata. Provider configuration, endpoints, and credentials are deliberately excluded.", "required": ["target_ref", "channel_id", "channel_version", "enabled", "revision"], "properties": {"target_ref": {"type": "string"}, "channel_id": {"type": "string"}, "channel_version": {"type": "string"}, "label": {"type": "string"}, "formats": {"type": "array", "items": {"type": "string"}}, "enabled": {"type": "boolean"}, "revision": {"type": "string"}}, "additionalProperties": false},
      "NotificationTargetListResponse": {"type": "object", "required": ["targets"], "properties": {"targets": {"type": "array", "items": {"$ref": "#/components/schemas/NotificationTargetView"}}}, "additionalProperties": false},
...
      "NotificationTargetCreateRequest": {"type": "object", "description": "Configuration is provider-owned opaque JSON. It is accepted only for a create mutation and never serialized in a response.", "required": ["target_ref", "channel_id", "channel_version", "config", "enabled"], "properties": {"tenant_id": {"type": "string"}, "target_ref": {"type": "string"}, "channel_id": {"type": "string"}, "channel_version": {"type": "string"}, "label": {"type": "string"}, "formats": {"type": "array", "items": {"type": "string"}}, "config": {"not": {"type": "null"}, "writeOnly": true, "description": "Opaque provider configuration; may contain secret material and is never returned."}, "enabled": {"type": "boolean"}}, "additionalProperties": false},
      "NotificationTargetUpdateRequest": {"type": "object", "description": "Configuration is optional on update. Omitting it preserves the existing provider configuration; supplying it replaces that configuration. It is never returned.", "required": ["target_ref", "channel_id", "channel_version", "enabled", "expected_revision"], "properties": {"tenant_id": {"type": "string"}, "target_ref": {"type": "string"}, "channel_id": {"type": "string"}, "channel_version": {"type": "string"}, "label": {"type": "string"}, "formats": {"type": "array", "items": {"type": "string"}}, "config": {"not": {"type": "null"}, "writeOnly": true, "description": "Optional opaque provider configuration; may contain secret material and is never returned."}, "enabled": {"type": "boolean"}, "expected_revision": {"type": "string"}}, "additionalProperties": false},
      "NotificationTargetDeleteRequest": {"type": "object", "required": ["target_ref", "expected_revision"], "properties": {"tenant_id": {"type": "string"}, "target_ref": {"type": "string"}, "expected_revision": {"type": "string"}}, "additionalProperties": false},
...
      "Account": {"type": "object", "required": ["account", "role", "tenant_id", "status", "must_change_password", "created_at"], "properties": {"account": {"type": "string"}, "email": {"type": "string"}, "role": {"type": "string", "enum": ["admin", "tenant_admin", "user"]}, "tenant_id": {"type": "string"}, "status": {"type": "string", "enum": ["pending_activation", "active", "disabled"]}, "must_change_password": {"type": "boolean"}, "created_at": {"type": "string", "format": "date-time"}}, "additionalProperties": false},
      "AccountList": {"type": "object", "required": ["accounts"], "properties": {"accounts": {"type": "array", "items": {"$ref": "#/components/schemas/Account"}}}, "additionalProperties": false},
      "AccountCreateRequest": {"type": "object", "required": ["account", "password", "role", "tenant_id"], "properties": {"account": {"type": "string"}, "email": {"type": "string"}, "password": {"type": "string", "minLength": 8, "maxLength": 72, "writeOnly": true}, "role": {"type": "string", "enum": ["admin", "tenant_admin", "user"]}, "tenant_id": {"type": "string"}}, "additionalProperties": true},
      "AccountCreateResponse": {"type": "object", "required": ["account", "email", "role", "tenant"], "properties": {"account": {"type": "string"}, "email": {"type": "string"}, "role": {"type": "string"}, "tenant": {"type": "string"}}, "additionalProperties": false},
      "PasswordSetRequest": {"type": "object", "required": ["password"], "properties": {"password": {"type": "string", "minLength": 8, "maxLength": 72, "writeOnly": true}}, "additionalProperties": true},
      "AccountStatusRequest": {"type": "object", "required": ["status"], "properties": {"status": {"type": "string", "enum": ["active", "disabled"]}}, "additionalProperties": true},
...
    }
  }
}